==> notebooks/03-nomic-exploration.ipynb <==
# Nomic Exploration

I'm going to use the Nomic SDK to create embeddings for the abstracts. I'll use the Nomic object to retrieve the 2D projection using their projection model, as well as the topic modeling features.

import pandas as pd
from nomic import atlas
from nomic import AtlasDataset

df = pd.read_csv("../data/biorxiv_complete_dataset.csv")

df.head()

df.shape

## Clean data

# Filter for only version 1 papers
df = df[df['version'] == 1]

df.shape

# Drop rows with missing abstracts
df = df.dropna(subset=['abstract'])

df.shape

df.head()

# Check for duplicate DOIs
duplicate_dois = df[df.duplicated(subset=['doi'], keep=False)]
print(f"Number of rows with duplicate DOIs: {len(duplicate_dois)}")

duplicate_dois

# Keep only first occurrence of each DOI
df = df.drop_duplicates(subset=['doi'], keep='first')


df.shape

df.head()

## Create and upload atlas

# This only runs once on first upload.
atlas = atlas.map_data(
  data=df, 
  indexed_field="abstract", 
  identifier="orchard", 
  is_public=False,
  id_field="doi"
)

atlas = AtlasDataset("orchard")

map = atlas.maps[0]
projected_embeddings = map.embeddings.projected

projected_embeddings.shape

projected_embeddings.head()

atlas.get_data(["10.1101/2023.06.28.546819"])

topic_df = map.topics.df

topic_df.head(20)

viral_infections_df = topic_df[topic_df['topic_depth_1'] == "Viral Infections (2)"]
viral_infections_df
