==> notebooks/04-create-db.ipynb <==
# Create a Full Database

I've now got all of the bioRxiv abstracts, embeddings, projections, and topics. I'll merge all of these together into one database.

import pandas as pd
from nomic import AtlasDataset

df = pd.read_csv("../data/biorxiv_complete_dataset.csv")

I'll perform all the same filtering as before.

df = df[df['version'] == 1]
df = df.dropna(subset=['abstract'])
df = df.drop_duplicates(subset=['doi'], keep='first')

df.shape

And then I'll get the atlas for the projections and topics.

atlas = AtlasDataset("orchard")
projections = atlas.maps[0].embeddings.projected

projections.head()

projections.shape

df.head()

merged_df = df.merge(projections, on='doi')
merged_df.head()

merged_df[merged_df['doi'] == '10.1101/2023.06.28.546819']

topic_df = atlas.maps[0].topics.df

topic_df.head()

merged_df = merged_df.merge(topic_df, on='doi')
merged_df.head()

unique_servers = merged_df['server'].unique()
unique_servers


merged_df['type'].unique()

merged_df = merged_df.drop(columns=['server', 'version', 'jatsxml', 'license', 'type'])

merged_df.head()

merged_df.to_csv('../data/biorxiv_nomic.csv', index=False)

unique_topic_depth_1 = merged_df['topic_depth_1'].unique()
unique_topic_depth_1

unique_topic_depth_2 = merged_df['topic_depth_2'].unique()
unique_topic_depth_2

len(unique_topic_depth_2)

unique_topic_depth_3 = merged_df['topic_depth_3'].unique()
unique_topic_depth_3

len(unique_topic_depth_3)

len(unique_topic_depth_1)

aging_topics_2 = [topic for topic in unique_topic_depth_2 if "Aging" in topic]
aging_topics_2


aging_topics_3 = [topic for topic in unique_topic_depth_3 if "Aging" in topic]
aging_topics_3


import re

def trim_topic(topic):
    return re.sub(r' \(\d+\)$', '', topic)

merged_df['topic_depth_1'] = merged_df['topic_depth_1'].apply(trim_topic)
merged_df['topic_depth_2'] = merged_df['topic_depth_2'].apply(trim_topic)
merged_df['topic_depth_3'] = merged_df['topic_depth_3'].apply(trim_topic)

merged_df['topic_depth_1'].unique()

merged_df.to_csv('../data/biorxiv_nomic.csv', index=False)

merged_df['topic_depth_2'].unique()

len(merged_df['topic_depth_2'].unique())
