==> notebooks/05-normalize-xy.ipynb <==
# Normalize x and y coordinates

The [regl-scatterplot](https://github.com/flekschas/regl-scatterplot/tree/main) library uses a normalized coordinate system. This means that the x and y coordinates are scaled to a range of -1 to 1.

I need to convert the (x,y) coordinates to this normalized coordinate system, because it's too slow and and memory intensive to do this in the browser.



import pandas as pd

df = pd.read_csv('../data/biorxiv_nomic.csv')

df.head()

# Normalize the x and y values to be between -1 and 1
x_min, x_max = df['x'].min(), df['x'].max()
y_min, y_max = df['y'].min(), df['y'].max()

df['x'] = 2 * (df['x'] - x_min) / (x_max - x_min) - 1
df['y'] = 2 * (df['y'] - y_min) / (y_max - y_min) - 1

df.head()

df.to_csv('../data/biorxiv_nomic_normalized.csv', index=False)

Steps for uploading to Neon ([docs](https://neon.tech/docs/import/import-from-csv)):

1. `psql [db-credentials]`
2. `TRUNCATE public.orchard` (Delete previous data)
3. `\copy public.orchard FROM './data/biorxiv_nomic_normalized.csv' DELIMITER ',' CSV HEADER`
