==> notebooks/06-create-db-index.ipynb <==
# Create DB Index

I need to have a numberical index column for the db. That way, I can select the rows that are selected in the scatterplot in the oRchard app.

import pandas as pd

df = pd.read_csv('../data/biorxiv_nomic_normalized.csv')

df.head()

df['id'] = df.index

df.to_csv('../data/biorxiv_nomic_normalized.csv', index=False)
